AutoGen/requirements-dev.txt
autogen-ext[video-surfer , web-surfer,task-centric-memory,ollama , openai,file-surfer]
autogen-agentchat[gemini]